VER3-124 factored out Hero and added HeroDetailComponent
