+++ src/app/app.component.ts
@@ -1,9 +1,5 @@
 import { Component } from '@angular/core';
-
-export class Hero {
-  id: number;
-  name: string;
-}
+import {Hero} from "./hero";
 
 const HEROES: Hero[] = [
   { id: 11, name: 'Mr. Nice'},
@@ -28,15 +24,7 @@ const HEROES: Hero[] = [
         [class.selected]="hero === selectedHero">
       <span class="badge">{{hero.id}}</span> {{hero.name}}
     </li>
-  </ul>
-  <div *ngIf="selectedHero">
-    <h2>{{selectedHero.name}} details!</h2>
-    <div><label>id: </label>{{selectedHero.id}}</div>
-    <div>
-      <label>name: </label>
-      <input [(ngModel)]="selectedHero.name" placeholder="name">
-    </div>
-  </div>` ,
+  </ul>` ,
   styles: [`
     .selected {
       background-color: #CFD8DC !important;
