+++ src/app/hero.ts
@@ -1,0 +1,7 @@
+/**
+ * Created by alasdair.macleod on 6/28/2017.
+ */
+export class Hero {
+  id: number;
+  name: string;
+}
